requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
pydantic
pyyaml
joblib
python-dateutil
requests